% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnvset}
\alias{cnvset}
\title{Copy Number Variation (CNV) data}
\format{A data matrix with 534 genes copy change and 3 samples}
\usage{
cnvset
}
\description{
A dataset containing the genome-wide gene copy change identified in the 3 multiple myeloma cell line samples.
The variables are as follows
}
\details{
\itemize{
 \item gene
 \item copy number variation segment ID
 \item sample 1 copy change segment mean value
 \item sample 2 copy change segment mean value
 \item sample 3 copy change segment mean value
 }
 @source \url{https://research.themmrf.org/
 }
}
\keyword{datasets}
