% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePS.R
\name{computePS}
\alias{computePS}
\title{Computes the profile statistics}
\usage{
computePS(rd1, cd1, cd2, profile)
}
\arguments{
<<<<<<< HEAD
\item{rd1}{: A numeric value of the reference sample (R) on which to estimate the profile statistics for a given row}
=======
\item{rd1}{: A numeric value of the reference sample (R) on which to estimate the profile statistics for a given gene, gene probe or gene copy segment}
>>>>>>> upstream/master

\item{cd1}{: A numeric value of the comparison sample 1 (S1) for a given gene, gene probe or gene copy segment}

\item{cd2}{: A numeric value of the comparison sample 2 (S2) for a given gene, gene probe or gene copy segment}

\item{profile}{: The desired direction of genomic change. The values are "up" (default) or "down" to select for increased or decreased gene set profile, respectively}
}
\value{
The returned value is profile statistics computed considering the specified change in the reference sample when compared to the remaining two relative samples.
}
\description{
Computes the increased or decreased profile statistics for each row (or gene) across the three samples within a feature or data type (expression, methylation, or copy-number variation)
}
\details{
This function requires three data values corrosponding to three samples for a given gene (or row), respectively
}
\examples{
rd1 = 40
cd1 = 20
cd2 = 20
computePS(rd1, cd1, cd2, profile="up")
}
\author{
Bhakti Dwivedi & Jeanne Kowalski
}
\keyword{Profile}
\keyword{Statistics}
