% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptModel.R
\name{cptModel}
\alias{cptModel}
\title{Computes within and between gene feature profile statistics by feature and amongst features}
\usage{
cptModel(psm, genelist, cpt.data, cpt.method, cpt.max)
}
\arguments{
\item{psm}{: A data matrix of estimated gene profile statistics for each feature}

\item{genelist}{: A vector of gene names or gene symbols corrosponding to the profile statistics}

\item{cpt.data}{: Identify changepoints in the data using variance (cpt.var), mean (cpt.mean) or both (meanvar). Default is cpt.var.}

\item{cpt.method}{: Choice of single or multiple changepoint model. Default is "BinSeg".}

\item{cpt.max}{: The maximum number of changepoints to search for using "BinSeg" method. Default is 60. This number is dependent on the number of input data points}
}
\value{
Estimated change points in the input data set
}
\description{
Computes the percentiles on the estimated profile statistics within a gene and across genes for one or more combination of feature or data types (expression, methylation, copy-number variation, or variant change)
}
\details{
This function estimates within and between feature profile statistics by gens in addition to the summed percentiles and successive differences
}
\examples{
id <- 1000 ## number of probes
s <- 3 ## number of sample groups
dm <- matrix(runif(id*s,0,200), nrow=id, ncol=s, dimnames=list(paste("gene", 1:id, sep="") , paste("fs", 1:s, sep="")))
genelist <- rownames(dm)
cptModel(dm, genelist, cpt.data="var", cpt.method="BinSeg", cpt.max=60)
}
\author{
Bhakti Dwivedi & Jeanne Kowalski
}
\keyword{Profile}
\keyword{Statistics}
