% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptPlot.R
\name{cptPlot}
\alias{cptPlot}
\title{Scatterplot representation of gene sets by change points}
\usage{
cptPlot(psv, cut.pts)
}
\arguments{
\item{psv}{: A data vector of estimated profile statistics on which changepoints are identified}

\item{cut.pts}{: The estimated profile statistics cutoffs corrosponding to the locations in psv}
}
\value{
Plot representing all the identified gene sets by change points in the data
}
\description{
Scatterplot representation of identified change points on the estimated profile statistics within the data
}
\details{
This function expects 'gispa.output' profile statistics output from GISPA.R main function
}
\examples{
x <- runif(100, 0.0, 1.0)
y <- c(0.2, 0.6, 0.8)
cpt.plot <- cptPlot(psv=x,cut.pts=y)
}
\author{
Bhakti Dwivedi & Jeanne Kowalski
}
