% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptSlopeplot.R
\name{cptSlopeplot}
\alias{cptSlopeplot}
\title{Scatterplot representation of identified change points gene set slopes}
\usage{
cptSlopeplot(gispa.output,feature,type)
}
\arguments{
\item{gispa.output}{: A data matrix of between gene feature profile statistics for each feature with corrosponding identified changepoints. The row names should corrospond to genes or names to be displayed on y-axis}

\item{feature}{: Analysis type i.e., one ('1'), two ('2') or three ('3') dimensional feature analysis.}

\item{type}{: Type of data, e.g., EXP  (default) for expression, VAR of variants, CNV for copy number change.}
}
\value{
Scatterplot illustrating the average slopes by change point to access the best gene set profile
}
\description{
This function will plot the average slopes estimated over all gene sets within each change point by data types
}
\details{
This function expects the output from GISPA function of GISPA package, and highlights the gene set slope profile in the selected changepoints
}
\examples{
id <- 200 ## number of rows
s <- 3 ## number of columns
dm <- matrix(runif(id*s,0,10), nrow=id, ncol=s, 
                 dimnames=list(paste("gene", 1:id, sep=""), 
                 paste("sample", 1:s, sep="")))
changepoints <- sort(sample(1:2, id, replace=TRUE))
dm <- cbind(dm,changepoints)
cptSlopeplot(gispa.output=dm,feature=1,type="EXP")
}
\author{
Bhakti Dwivedi & Jeanne Kowalski
}
