% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprset}
\alias{exprset}
\title{Gene Variant (EXP) data}
\format{
A data matrix with 1500 genes and 3 samples
}
\usage{
exprset
}
\description{
A dataset containing the genome-wide gene expression values from 3 multiple myeloma cell line samples.
The variables are as follows
}
\details{
\itemize{
 \item gene
 \item gene names
 \item sample 1 log2 transformed normalized expression count values
 \item sample 2 log2 transformed normalized expression count values
 \item sample 3 log2 transformed normalized expression count values
 }
 @source \url{https://research.themmrf.org/
 }
}
\keyword{datasets}
