% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISPA.R
\name{GISPA}
\alias{GISPA}
\title{Gene Integrates Set Profile Analysis}
\usage{
GISPA(feature,f.sets,g.set,ref.samp.idx,comp.samp.idx,f.profiles,cpt.data,cpt.method,cpt.max)
}
\arguments{
\item{feature}{: Analysis type i.e., one ('1'), two ('2') or three ('3') dimensional feature analysis.}

\item{f.sets}{: A list of ExpressionSet data objects corrosponding to a data type}

\item{g.set}{: A GeneSet from an ExpressionSet to subset the f.sets for analysis purposes. 'geneIds' should corrospond to the gene names. Default is null, i.e., genome-wide analysis}

\item{ref.samp.idx}{: Reference sample column index on which to determine the gene set profile. The default is 3}

\item{comp.samp.idx}{: The other two relative sample column index against which the profile is being determined. The default is 4 and 5}

\item{f.profiles}{: A vector of the desired direction of genomic change (or profile) corrosponding to each data type. The values are "up" or "down" for increased and decreased gene set profile, respectively}

\item{cpt.data}{: Identify changepoints for data using variance (cpt.var), mean (cpt.mean) or both (meanvar). Default is cpt.var.}

\item{cpt.method}{: Choice of single or multiple changepoint model. Default is "BinSeg".}

\item{cpt.max}{: The maximum number of changepoints  to search for using "BinSeg" method. Default is 60. This number is dependent on the number of input data points}
}
\value{
The returned value is a data matrix including the original data along with between gene profile statistics and identified changepoints.
}
\description{
Identifies gene sets with a similar a prior defined profile using any combination of three feature or data types
}
