% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propBarplot.R
\name{propBarplot}
\alias{propBarplot}
\title{A plotting function for proportion by sample}
\usage{
propBarplot(gispa.output,feature,cpt,input.cex,input.cex.lab,ft.col,strip.col)
}
\arguments{
\item{gispa.output}{: A data matrix of between gene feature profile statistics for each feature with corrosponding identified changepoints. The row names should corrospond to genes or names to be displayed on y-axis}

\item{feature}{: Analysis type i.e., one ('1'), two ('2') or three ('3') dimensional feature analysis.}

\item{cpt}{: Changepoints to be plotted.}

\item{input.cex}{: character (or symbol expansion) for the x- and y-axis labels}

\item{input.cex.lab}{: character (or symbol expansion) for the horizontal and vertial strip labels}

\item{ft.col}{: a vector of colors of the bar for the features or data types}

\item{strip.col}{: color to be used for the vertial strip}
}
\value{
Barplot illustrating each sample proportion for each gene in the selected change point
}
\description{
Given a gene, this function will plot the proportion of each sample over the three samples within each data type
}
\details{
This function expects the output from the main function of GISPA package, and highlights the gene set in the selected changepoints and their proportion in each of the three sample groups by data type.
}
\examples{
id <- 20 ## number of rows
s <- 3 ## number of columns
dm <- matrix(runif(id*s,0,10), nrow=id, ncol=s, 
                 dimnames=list(paste("gene", 1:id, sep=""), 
                 paste("fs", 1:s, sep="")))
changepoints <- sort(sample(1:2, id, replace=TRUE))
dm <- cbind(dm,changepoints)
propBarplot(gispa.output=dm,feature=2,cpt=1,
           input.cex=0.5,input.cex.lab=0.5,
           ft.col=c("grey0", "grey60"),strip.col="yellow")
}
\author{
Bhakti Dwivedi & Jeanne Kowalski
}
