% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varset}
\alias{varset}
\title{Variant (VAR) data}
\format{A data matrix with 1101 genes variants and 3 samples}
\usage{
varset
}
\description{
A dataset containing the genome-wide gene variant proportion from 3 multiple myeloma cell line samples.
The variables are as follows
}
\details{
\itemize{
 \item gene
 \item gene names
 \item sample 1 transformed variant proportion data
 \item sample 2 transformed variant proportion data
 \item sample 3 transformed variant proportion data
 }
 @source \url{https://research.themmrf.org/
 }
}
\keyword{datasets}
